move evaluate_state back into original file
